test(ci): add ci
